+++ packages/footer/src/styledComponents/HelperContent.js
@@ -1,5 +1,4 @@
 import styled from "styled-components";
-import { colors as colorConstants, fontWeights } from "@crave/farmblocks-theme";
 
 const HelperContentContainer = styled.div`
   font-family: Lato, sans-serif;
@@ -16,16 +15,6 @@ const HelperContentContainer = styled.div`
 
   > .image {
     margin-right: 16px;
-  }
-
-  > .link {
-    color: ${colorConstants.OYSTER};
-    font-weight: ${fontWeights.SEMIBOLD};
-    text-decoration: none;
-
-    :hover {
-      color: ${colorConstants.INDIGO_MILK_CAP};
-    }
   }
 `;